moved skipspeedtestjob to ookla namespace
