+++ app/Jobs/Ookla/ProcessSpeedtestBatch.php
@@ -4,7 +4,6 @@ namespace App\Jobs\Ookla;
 
 use App\Jobs\CheckAndUpdateBenchmarkResult;
 use App\Jobs\CheckForInternetConnectionJob;
-use App\Jobs\SkipSpeedtestJob;
 use App\Models\Result;
 use Illuminate\Bus\Batch;
 use Illuminate\Contracts\Queue\ShouldQueue;
